+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - User option to specify minimum width for a single breadcrumb entry ([#30](https://github.com/benji300/joplin-note-tabs/issues/30))
+- User option to specify the font size for the panel
 - Command to toggle panel visibility ([#16](https://github.com/benji300/joplin-note-tabs/issues/16))
 
 ### Changed